requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0

# RAG System
rank-bm25>=0.2.2
jieba>=0.42.1
chromadb>=0.4.0

# Vector & Embedding
sentence-transformers>=2.2.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pydantic>=2.0.0
pandas>=2.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.23.0

# Utils
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development (Optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
